pyproject.toml
[tool.poetry]
name = "brown-globus-utils"
version = "0.1.0"
description = ""
authors = ["Isabel Restrepo <isabelrestre@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
globus_sdk = "^1.7"
python-dotenv = "^0.10.3"
argparse = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
pylint = "^2.3"

[tool.poetry.scripts]
submit_transfer = "brown_globus_utils.submit_transfer:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
